Added ItemPhoto delete method handler
