+++ server/api/itemPhoto/[id].delete.ts
@@ -1,0 +1,45 @@
+import { defineEventHandler, setResponseStatus } from "h3";
+import type { User } from "../../../types/session";
+import { getServerSession } from "#auth";
+
+export default defineEventHandler(async (event) => {
+    const id = getRouterParam(event, 'id');
+    const prisma = event.context.prisma;
+    const session = await getServerSession(event);
+    const user = session?.user as User | undefined;
+
+    if (!user?.role || (user.role !== "SUPER" && user.role !== "ADMIN")) {
+        throw createError({
+            statusMessage: "Unauthenticated",
+            statusCode: 403,
+        });
+    }
+
+    if (!id) {
+        setResponseStatus(event, 400);
+        return {
+            success: false,
+            error: "ItemPhoto ID is required",
+        };
+    }
+
+    try {
+        const itemPhoto = prisma.itemPhoto.delete({
+            where: {
+                id: id,
+            }
+        });
+        setResponseStatus(event, 200);
+        return {
+            success: true,
+            itemPhoto: itemPhoto,
+        };
+    } catch (error) {
+        const errorMessage = error instanceof Error ? error.message : "Unknown error occurred";
+        setResponseStatus(event, 500);
+        return {
+            success: false,
+            error: `Error creating event: ${errorMessage}`,
+        };
+    }
+});
